+++ .gitignore
@@ -1,13 +1,14 @@
-.env
+__pycache__
+-pytest_cache
 env
+htmlcov
+/tests/__pycache__
+.coverage
+.env
 token_1.py
 /models/__pycache__
 /router/__pycache__
 /routers/__pycache__
-__pycache__
-.coverage
-/tests/__pycache__
 /models/__init__.py
 /routers/__init__.py
-htmlcov
 .pytest_cache
